refactor(agent): sever the @tools/registry cycle at the remote-agent client
`@tools/registry` imports every registered tool, so anything that reaches it
installs the LaTeX, Lean, arxiv and Zotero tool set. 19 of the 50 `defineTool()`
modules reached it — through one gateway — and therefore shared an identical
630-file module closure:

    bash -> @agent/storage -> @agent/index/agentRegistry
         -> remoteAgentMeta -> RemoteAgentLoader -> @tools/registry

`RemoteAgentLoader` had two unrelated halves. `listRemoteAgents()` reads
metadata rows and needs no tool resolution; `loadRemoteAgent()` parses agent
YAML and genuinely needs `resolveToolDefinitions`. The agent index only ever
wanted the first, but imported the module that carried both.

Split them: the listing half moves to `@agent/remote/remoteAgentList`, which has
no `@tools/registry` (or `@agent/index`) import, and `remoteAgentMeta` points at
it. `RemoteAgentLoader` keeps config loading and its single caller, `agentLoad`;
nothing under `src/tools/` reaches it any more. The static-only class collapses
to the two plain functions it always was.

bash's closure: 630 files / 34 domain files -> 217 / 0. 17 of the 19 are severed
the same way. `DelegationTools` and `WorkflowScriptTool` keep the registry on
purpose — they launch a subagent in-band, and resolving that subagent's YAML
tool list is a real dependency on the whole tool table, not a cycle.

Note that a lazy `await import()` would not have helped: dynamic-import targets
are still closure members (an npm consumer installs them), which is why the
edge is deleted rather than deferred.

`toolRegistryCycle.vitest.ts` pins this with the issue's own measurement — an
esbuild metafile bundle per `defineTool()` module, dependencies external. No
ratchet baseline: the assertion is absolute.

Closes #9327
